Gradle configuration are done. Now we need to run gradle clean fatJar to create an executable jar
