+++ src/main/java/com/dc/peer/InitConfig.java
@@ -1,3 +1,5 @@
+package com.dc.peer;
+
 import java.util.Arrays;
 
 public class InitConfig {
