+++ src/main/java/com/dc/peer/Node.java
@@ -1,10 +1,28 @@
+package com.dc.peer;
+
 class Node {
-    public String ip;
-    public int port;
-    public String userName;
+    private String ip;
+    private int port;
+    private String userName;
 
     public Node(String ip, int port) {
         this.ip = ip;
+        this.port = port;
+    }
+
+    public String getIp() {
+        return ip;
+    }
+
+    public void setIp(String ip) {
+        this.ip = ip;
+    }
+
+    public int getPort() {
+        return port;
+    }
+
+    public void setPort(int port) {
         this.port = port;
     }
