+++ src/main/java/com/dc/peer/StartNode.java
@@ -1,10 +1,12 @@
+package com.dc.peer;
+
 import java.util.Scanner;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class StartNode {
     public static void main(String[] args) {
         int port = ThreadLocalRandom.current().nextInt(1050, 50000);
-        PeerNodeNew1 node = new PeerNodeNew1("127.0.0.1",port,"node" + port);
+        PeerNode node = new PeerNode("127.0.0.1",port,"node" + port);
         Scanner in = new Scanner(System.in);
         while(true) {
             System.out.println("What you want to do?");
