+++ src/main/java/com/dc/peer/model/Comment.java
@@ -1,4 +1,4 @@
-package model;
+package com.dc.peer.model;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
