+++ src/main/java/com/dc/peer/model/Rank.java
@@ -1,4 +1,4 @@
-package model;
+package com.dc.peer.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
